pyproject.toml
[tool.poetry]
name = "maslib"
version = "0.1.0"
description = "Movie and Series Library"
authors = ["LrnzdDC3696 <lrnzdc3696@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
python-dotenv = "^0.21.0"
djangorestframework = "^3.14.0"
coverage = "^7.2.7"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
